move LoopbackPeer to test subdir
