+++ src/overlay/test/ItemFetcherTests.cpp
@@ -9,8 +9,8 @@
 #include "lib/catch.hpp"
 #include "main/ApplicationImpl.h"
 #include "overlay/ItemFetcher.h"
-#include "overlay/LoopbackPeer.h"
 #include "overlay/OverlayManager.h"
+#include "overlay/test/LoopbackPeer.h"
 #include "test/TestUtils.h"
 #include "test/test.h"
 #include "xdr/Stellar-types.h"
