+++ src/overlay/test/LoadManagerTests.cpp
@@ -2,8 +2,8 @@
 // under the Apache License, Version 2.0. See the COPYING file at the root
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
-#include "overlay/LoopbackPeer.h"
 #include "overlay/OverlayManager.h"
+#include "overlay/test/LoopbackPeer.h"
 #include "test/TestUtils.h"
 #include "test/test.h"
 #include "util/Timer.h"
