+++ src/overlay/test/LoopbackPeer.cpp
@@ -2,7 +2,7 @@
 // under the Apache License, Version 2.0. See the COPYING file at the root
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
-#include "overlay/LoopbackPeer.h"
+#include "overlay/test/LoopbackPeer.h"
 #include "crypto/Random.h"
 #include "main/Application.h"
 #include "medida/timer.h"
