+++ src/overlay/test/OverlayTests.cpp
@@ -8,10 +8,10 @@
 #include "main/Application.h"
 #include "main/Config.h"
 #include "overlay/BanManager.h"
-#include "overlay/LoopbackPeer.h"
 #include "overlay/OverlayManagerImpl.h"
 #include "overlay/PeerManager.h"
 #include "overlay/TCPPeer.h"
+#include "overlay/test/LoopbackPeer.h"
 #include "simulation/Simulation.h"
 #include "test/TestUtils.h"
 #include "test/test.h"
