+++ src/simulation/Simulation.h
@@ -9,8 +9,8 @@
 #include "main/Application.h"
 #include "main/Config.h"
 #include "medida/medida.h"
-#include "overlay/LoopbackPeer.h"
 #include "overlay/StellarXDR.h"
+#include "overlay/test/LoopbackPeer.h"
 #include "simulation/LoadGenerator.h"
 #include "test/TxTests.h"
 #include "util/Timer.h"
