+++ src/test/fuzz.cpp
@@ -8,9 +8,9 @@
 #include "main/Application.h"
 #include "main/Config.h"
 #include "main/StellarCoreVersion.h"
-#include "overlay/LoopbackPeer.h"
 #include "overlay/OverlayManager.h"
 #include "overlay/TCPPeer.h"
+#include "overlay/test/LoopbackPeer.h"
 #include "test/test.h"
 #include "util/Fs.h"
 #include "util/Logging.h"
