+++ src/transactions/test/BumpSequenceTests.cpp
@@ -6,7 +6,7 @@
 #include "lib/catch.hpp"
 #include "main/Application.h"
 #include "main/Config.h"
-#include "overlay/LoopbackPeer.h"
+#include "overlay/test/LoopbackPeer.h"
 #include "test/TestAccount.h"
 #include "test/TestExceptions.h"
 #include "test/TestUtils.h"
